Improve User Functionality code

+++ src/main/java/com/christos/app/twittercloneapi/services/UserService.java
@@ -1,7 +1,10 @@
 package com.christos.app.twittercloneapi.services;
 
+import com.christos.app.twittercloneapi.exceptions.exceptions.AlreadyExistsException;
+import com.christos.app.twittercloneapi.exceptions.exceptions.UserNotFoundException;
 import com.christos.app.twittercloneapi.models.User;
 import com.christos.app.twittercloneapi.repositories.UserRepository;
+import com.christos.app.twittercloneapi.utils.UpdateJsonUtils;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.BeanWrapper;
@@ -26,19 +29,47 @@ public class UserService {
     }
 
     public User getUserById(Long id) {
-        return userRepository.getUserById(id);
+        User user = userRepository.getUserById(id);
+        if (user == null){
+            throw new UserNotFoundException("There is no user with id: " + id);
+        }
+
+        return user;
     }
 
-    public void updateOrSaveUser(User user) {
+    public User updateUser(Long id, User updatedUser) {
+        User user = userRepository.getUserById(id);
+
+        if (user == null) {
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
+
+        UpdateJsonUtils.copyNonNullProperties(updatedUser, user);
         userRepository.save(user);
+
+        return user;
     }
 
-    public void deleteUser(User user) {
+    public User createUser(User newUser) {
+        User user = userRepository.getUserByUsernameOrEmail(newUser.getUsername(), newUser.getEmail());
+
+        if (user != null) {
+            throw new AlreadyExistsException("User already exists");
+        }
+
+        userRepository.save(newUser);
+        return newUser;
+    }
+
+    public User deleteUser(Long id) {
+        User user = userRepository.getUserById(id);
+
+        if (user == null) {
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
+
         userRepository.delete(user);
-    }
-
-    public User getUserByEmailOrUsername(String username, String email) {
-        return userRepository.getUserByUsernameOrEmail(username, email);
+        return user;
     }
 
 }
